DPDKLAB-189: Add contact policy for environments
Add the ability for vendors to control who receives patch performance
failure e-mails for their environments.

This may appear to be overkill based on the discussion at the last
meeting, where the consensus was to have the report e-mails be a reply
to the patch submission, but having these finer-grained controls allows
for easier testing and development by turning off most of the external
e-mail reporting. These controls may also be useful to vendors.