DPDKLAB-195: Update database schema for private patchsets
Move the is_official parameter to the PatchSet model and rename it to
is_public; it will be set to private if the patchworks_id is not
provided (it was previously required). This paves the way to allowing a
vendor to submit a private patchset for experimental work to be run only
on their environment.