DPDKLAB-157: Add clone operation to environments
The clone operation allows creating a modifiable copy of an environment
in order to track changes to an environment over time.

The intention is that an environment becomes read only after it is
cloned, or when test results are added to it. In the latter case, the
environment should be cloned and edits made to the clone, in order to
differentiate results from different "generations" of an environment.
This behavior will be implemented in future commits.