DPDKLAB-204: Allow blank values for null-able fields
Fields which are nullable should generally also have blank=True so that
the user can supply a blank value; otherwise, there is no point to
allowing them to be NULL in the database. Fix this for some existing
fields.

We do not touch the Environment owner field because going forward this
will be made non-nullable.