DPDKLAB-204: Fix relationship between contact policy and environment
Reverse the direction of the OneToOneField, which is needed in order for
the admin interface to present the contact policy as an inline field in
the model.

Unfortunately, this makes the raw Python code for creating an
environment slightly more annoying as the environment must be created
and then the measurement must be created immediately afterward, instead
of creating both at the same time. However, this change makes the admin
interface actually usable for creating an environment which is important
for getting off the ground.