DPDKLAB-185: Add custom update operation for environments and test runs
Add an update operation to the environment serializer to avoid a 500
error if the user attempts a PUT operation on an existing environment.
This is an edge case that will probably never be used but the code to
implement it is (theoretically) in place with this commit. Add a bunch
of tests to make sure this functionality works (or fails) properly in
all cases.