DPDKLAB-141: Fix custom filter names in HTML output
The django-filter Filter classes attempt to look up the corresponding
model fields to determine the "verbose" name to use on the HTML form. If
there is no corresponding field name, as is the case for our custom
"complete" and "has_tarball" filters, the filter name shows up on the
HTML form as "[invalid name]". Fix this by supplying the label attribute
to the BooleanFilter constructor which overrides this lookup and fixes
the issue.