DPDKLAB-248: Add pw_is_active field to Patch model
Add a field to the model to determine whether or not the patch in
Patchwork is still active. This will be used to determine whether or not
to continue testing the patchset, by marking any patchsets that have
inactive patches as incomplete, in turn causing them to be skipped when
looking for the next patchset.