DPDKLAB-265: Display overall result
Show the overall result using the same color formatting as in the table.
Use background color instead of foreground because the yellow color used
for the Patch Apply error is incredibly hard to see for a lot of people
if it is used as a text color on a white or off-white background.