DPDKLAB-281: Add user profile and subscription models
Add a UserProfile model which will be used to add user attributes that
are not part of the default User model. Because we did not override the
User model from the start, we are stuck using this extra layer of
indirection to add attributes to a user. Add a signal to create a
profile for every new user, and create profiles for every existing user
as part of the database migration.

Using our UserProfile model, add a Subscription model to represent the
association between a user and an environment. All e-mail preferences
will be part of this association, allowing users to eventually select
their e-mail preferences on a per-device basis. The system enforces that
users may only subscribe to their company's environments, and handles
signals appropriately to remove subscriptions as group memberships
change.