DPDKLAB-300: Fix unit tests
Use LiveServerTestCase to allow logging into through the API. Change the
login process to go through the custom login view. Since this is a
TransactionTestCase, we need to create uses before every test as opposed
to setting up the data before hand.
Added a workaround to a bug which we think is part of Django (Guardian
or REST) to avoid IntegrityError foreign key issues by clearing the
cache during the tearDown.