DPDKLAB-327: Handle 401 or 403 error fetching environments
If the user is not logged in, they are not allowed to fetch the list of
active environments in the API. This raises an exception in the caller
and results in a 500 error. Fix this by handling these status codes by
assuming an empty list (which is exactly what we want).