DPDKLAB-278: Add dashboard 500 Internal Server Error template
Add a template which renders a human-readable message that the backend
database is inaccessible. This template retains the navbar and footer
elements so that the user can still easily retry the request or access
the static links in the footer, but is otherwise static as required by
the default Django 500 error view.