DPDKLAB-325: Add live_since field to Environment model
Add a field to mark when an environment should be considered live, and
thus its results should affect the overall result shown on the front
page of the dashboard. For the purposes of the migration, make the
default value for existing instances the beginning of UNIX time
(1970-01-01), but then reset the default to null for new instances.