DPDKLAB-343: Enable caching for results
Add private caching to the dashboard home page a patchset page.
This works by caching the page every 10 minutes by:

1. (if available) using the server's cache as defined by the CACHE
settings variable.
2. Caching on the users browser by setting private=True. This makes it
so when a user is logged in, the cache is only saved locally, and not
on the server to avoid any possible data leakage.

With some testing between firefox and chrome, Django seems to be
smart about handling these caches with anonymous users, where it will
cache the page on the server (if 1 is active) even if private=True
and vary_on_cookie is set. So future anonymous users will benefit
from the cache.