DPDKLAB-368: Add log_upload_file field that stores actual file
It is much easier to present the log file via the dashboard if Django is
in charge of the file storage than if the file storage is done via
Jenkins artifacts, because we can continue to use the existing backend
session authentication method.

Keep both fields at the same time because the only way to migrate them
would be to have a data migration that downloaded files, which starts to
exit the scope of what migrations should do.