DPDKLAB-279: Update website to rely on patchworks
This involved deleted several fields in PatchSet and deleting the
Patches table.
- Add pw_get method which caches the responses of patchworks.
- Remove pagination since patchworks does not support pagination for the
events api. We can add pagination later just to avoid querying excess
number of series.
- Add request_mapping.json, which maps urls to json responses. This is
meant for development purposes so that we don't need to query the
patchwork API all the time during development and not need to set up
programmatic caching.
- Add ENVIRONMENT variable, which is only used for above. I wanted to
avoid using the DEBUG setting because sometimes we take off DEBUG for
testing error pages.
- Add logging to dev_defaults.
- patchwork_range_str has been moved from the results to the dashboard
since it's based on the series.

DPDKLAB-341: Add completed timestamp to use with comparing our database
with patchworks api for what needs to be added to our database.

DPDKLAB-341: The page subject now uses the series title.

DPDKLAB-387: With the new API change, several URLs have been updated to
avoid 301s