DPDKLAB-399: Add disclaimer to the bottom of page
This was requested by intel. This only shows if any result shows up.
This is mostly a quick feature addition that does not modify the
database. If we get more disclaimers in the future, we may think about
adding them to the database, and only show based on whats on the page.