DPDKLAB-409: Log our own exceptions
Since the library picked did not have any global methods to log our own
errors, I "forked" the project to update that part. The only difference
is that the js methods are not wrapped in their own scope. (I also added
the license from the original repository, and added a readme describing
the changes.)