DPDKLAB-348: verify test_run download not public
This makes it so that the test run download is never public and can only
be accessed by the group responsible for the environment.

This does add a "bait and switch" to the dashboard, where the download
button shows up, but clicking it as an anonymous user leads to the
private storage page. I would like to remove the button, but that might
come at a later date.