patchset list: load table rows with javascript
Instead of waiting for the table to initially load (~5-10 seconds
on production), load the page without the table, and use javascript to
fetch individual rows of the table. We could fetch the whole table, but
there were some timeout issues when fetching the whole table. So
instead, fetch each row individually. This will help with avoiding
timeouts, and later on, we can implement caching for these individual
requests.

dashboard: add IE11 support

Since some users are using IE11, we need to transpile the javascript so
that it runs properly in IE11. We also need several polyfills. Thus, we
now have an NPM build system for our Javascript files. The js files were
moved to their own js files, and the npm build command will transpile
them into the static/build directory, allowing collectstatic to
function.