test run: add baseline information
https://bugs.dpdk.org/show_bug.cgi?id=195

This adds a baseline to the testrun model for when reruns are ran on a
different baseline. This removes the baseline definition on the detailed view,
and replaced with "applied on".

- Merge applied on and baseline to applied on in detailed view
- Move test time run at to be a dt element for consistency
- Add baseline information to test run if it exists
- Add branch to patchset model to follow tarball (and now testrun) models
- If commit_id is not set, return an empty string for consistency
- Removed the dpdk master check in the detailed results. This may be
  confusing once we implement subtree testing.

New test runs going forward will have a new baseline added. We decided to not
migrate all test run baselines to be the tarball applied on commit hash,
because existing baseline testruns would be pointing to themselves (since
baseline tarballs commit id, IS the commit they are on), when they are supposed
to point to their previous baseline.

For old patchset related test runs, assume that the baseline is same as the
"applied on" commit id.