tarballs: check if patchset exists before query
Since we now show a status on the tarball, it now gets called on the
tarball list, not just the patchset list. However, older tarballs dont
have a date and don't have a patchset. This will now make sure a
patchset and date exists before including it in the query.

- don't show "None" or empty if there is no date.
- set next url to current page instead of trying to reverse.