results: add indeterminate test result
Based on the following discussion:
Add a new status called "Indeterminate". This would be the
"fall through" case instead of using Possible Regression as
a fall through case. If an indeterminate result exists,
then the overall result (between all environments) will be
Indeterminate, unless there exists a Possible Regression,
in which case the overall result would be a Possible
Regression.