testrun: faster permission checking
Add a direct foreign key to test run for faster permission checking.
There was an n+1 type of issue with permission checks, where the id of
every applicable object was being returned when checking permissions for
a single object. Thus the API getting slower as more testruns were
added.

Since this is being added after the fact, this breaks existing public
permissions, thus they need to be set back to private and set back to
public for this to take effect.