env actions: run set private/public in threadpool
Since it takes a long time to run set public and set private, run it in
a thread.
Using a shared single threaded pool allows for a single worker to execute
set_public and set_private at a time.
Once this exists in the preferences view, this may use some type of
websocket and channel communication, but this is simple enough for now.
The initial page view has also been simplified instead of timing out
when trying to generate the page.

- Create decorator to handle uncaught exceptions in child threads.