testrun: Add testcase to test run
Since there are possible test harness errors, we may get no results for
a test run. Since there are no results, we are unable to determine what
test case the test run was meant for.
This adds a testcase key to the test run, so that when submitting
results, a testcase can be added.
Add filter to testcase to find the url when given a name when submitting
results.