subscriptions: fix public environments showing
Public environments would show up in the subscription page to users that
are not part of the group. This was already handled server side, where an
error was thrown to the user since they don't have permissions to add
subscriptions for the environment. This patch just makes it so that the
environments don't show up in the subscriptions page.