manage users: add user to group
Add the ability for a primary contact to add existing users to their
group

Permissions needed to be changed for this to work for the user viewset:
if POST:
    get all users.
else:
    like before, only get users in the group.
When a POST is done, there is extra permission checking, such as
verifying if the requesting user is a pc for any group. However, this is
not enough, because then they could add users to groups they are not
part of. To remedy this, an explicit permission check is done in the
manage-group action.