dashboard cache: use Django's cache
This cache method used to only be a "request cache", where for a single
dashboard request, it would cache API calls, then clear the cache.
Now this is only the case in a development environment.
This was fine for a while, but it was realized that all the requests
used thus far has been for unauthenticated urls and there were
essentially "cache misses" when we used javascript to load parts of the
page. Using a session cache was also not desirable, because the
information thus far could be shared with any users and session caches
last too long for this use case.