Add a generic way to build pipelines
Only admins can do this at the moment. This is meant for times where the
tests failed to run before the results could be uploaded to the
database. This removes needing to juggle between finding out the tarball
meta url and manually building the relevant pipelines for the
patchset/tarball.