Fix 0 being passed to subscription api
Selecting "inherit" in the subscription view, sends null (which is
expected). But we need to forward the request to the API server. Since
it gets received as a string, we convert it to a dict, which converts to
None. When hitting the API, None gets converted to 0 since the None
keyword does not exist in json. Instead send as json instead as data.