Separate result summary into separate request
In the production database, it takes about 2000 sql queries to list
results. This is because of the result summary. Removing it yields 100
sql queries. To reduce the initial page further, separate the query into
its own route/action, and get when needed.

Since getting the whole listing is now fast enough, it is now loaded
first, while each result summary is individually loaded.

With this new approach, the custom caching in the results is actually
worse, so it was removed.