Simplify result summary and remove incomplete
The incomplete status has been removed. This was because from a user
perspective, it was unclear what it meant. It was originally meant for a
single environment for a single test case. As test cases grew, it was no
longer 1 to 1, and some test cases did not apply to some environments.
Thus for environments with more than one test case, the incomplete
status was essentially meaningless. There are also environments that do
not run certain test cases. This could be resolved by adding a many
relation testcases to the environment, but that would just make the
queries more complicated, add more data to duplicate during cloning
environments, and possible add more load time to result summaries.
It was not really worth it for little gain.